gr-uhd/lib: support fmtlib 10
In rfnoc_rx_streamer_impl::work , d_metadata.error_code has
type uhd::rx_metadata_t::error_code_t , which is just enum.

But with fmtlib 10, fmt::format won't accept enum by default.
So just cast this variable error_code to int to make fmtlib-10
accept this.

Closes #6735 .

(cherry picked from commit b6c147dc1b78b706a2ebb6d78e5199f3ad0f8e83)

+++ gr-uhd/lib/rfnoc_rx_streamer_impl.cc
@@ -143,7 +143,7 @@ int rfnoc_rx_streamer_impl::work(int noutput_items,
     default:
         d_logger->warn("RFNoC Streamer block received error {:s} (Code: {})",
                        d_metadata.strerror(),
-                       d_metadata.error_code);
+                       static_cast<int>(d_metadata.error_code));
     }
 
     if (d_metadata.end_of_burst) {
